Allow multiple UTF8Readers to be used in the same thread at the same time
There's currently a bug where UTF8Readers that exist concurrently in the
same thread overwrite each other, because they improperly share buffers.
This is clearly a bug; freeBuffers is kinda a no-op.

This PR modifies UTF8Reader to avoid this.

When we take the buffer, we set it to null, and replace it when we free
the buffers. This means that people who use the class in the same
thread merely create a new buffer, which is a concern of performance and
not of correctness