Add new feature LITERAL_BLOCK_STYLE
If enabled will use a literal block style for string containing newlines. All other quotes are unaffected.
This make sense in addition to MINIMIZE_QUOTES and ALWAYS_QUOTE_NUMBER_AS_STRINGS as e.g this is safer for
string values like "true" (which when quotes are stripped mutate from a string type to a boolean type).

It is false by default.